check order to actions, promise on setUserOrder
