+++ src/components/RegistrationLiteComponent.js
@@ -15,7 +15,7 @@ import { FragmentParser } from "openstack-uicore-foundation/lib/components";
 import { doLogin, passwordlessStart } from 'openstack-uicore-foundation/lib/methods'
 import { getEnvVariable, SUMMIT_API_BASE_URL, OAUTH2_CLIENT_ID, REGISTRATION_BASE_URL } from '../utils/envVariables'
 
-import { getUserProfile, setPasswordlessLogin, setUserOrder, updateProfile } from "../actions/user-actions";
+import { getUserProfile, setPasswordlessLogin, setUserOrder, checkOrderData } from "../actions/user-actions";
 import { getThirdPartyProviders } from "../actions/base-actions";
 
 import 'summit-registration-lite/dist/index.css';
@@ -29,7 +29,7 @@ const RegistrationLiteComponent = ({
     getUserProfile,
     setPasswordlessLogin,
     setUserOrder,
-    updateProfile,
+    checkOrderData,
     loadingProfile,
     loadingIDP,
     summit,
@@ -56,19 +56,6 @@ const RegistrationLiteComponent = ({
     const onClickLogin = (provider) => {
         doLogin(getBackURL(), provider);
     };
-
-    const checkOrderData = (order) => {        
-        const { owner_company, owner_first_name, owner_last_name } = order;
-        const { company, given_name, family_name } = registrationProfile;
-        if (owner_company !== company || owner_first_name !== given_name || owner_last_name !== family_name) {
-            const newProfile = {
-                first_name: owner_first_name,
-                last_name: owner_last_name,
-                company: owner_company
-            };
-            updateProfile(newProfile);
-        }        
-    }
 
     const formatThirdPartyProviders = (providers_array) => {
         const providers = [
@@ -132,12 +119,11 @@ const RegistrationLiteComponent = ({
         },
         goToEvent: () => navigate('/a/'),
         goToRegistration: () => navigate(`${getEnvVariable(REGISTRATION_BASE_URL)}/a/${summit.slug}`),
-        onPurchaseComplete: async (order) => {
+        onPurchaseComplete: (order) => {
             // we are informing that we did a purchase recently to widget
             setIsRecentPurchase(true);
-            setUserOrder(order);
             // check if it's necesary to update profile
-            await checkOrderData(order);
+            setUserOrder(order).then(_ => checkOrderData(order));
         },
         inPersonDisclaimer: siteSettings?.registration_in_person_disclaimer
     };
@@ -176,5 +162,5 @@ export default connect(mapStateToProps, {
     getUserProfile,
     setPasswordlessLogin,
     setUserOrder,
-    updateProfile
+    checkOrderData
 })(RegistrationLiteComponent)
